Fix error when bundle file have directories
The .tar.gz bundle file that is downloaded from the server can now embbed
directories but it does not contains TAR header entries for
directories.

This lead to attempt to extract files to directories that are not
created.

This commit aims at doing a fallback: if a file fails to be created, we
first try to create it's directory, regardless of if there was or not a
TAR directory header.